handwritten digit recognition/requirements.txt
tensorflow
numpy
matplotlib
seaborn
scikit-learn
streamlit
streamlit-drawable-canvas
pillow
